requirements.txt
PyQt5
ffmpeg-python
transformers
nltk
openai-whisper
torch
python-dotenv
